Fix vim highlighting of tabs and extra whitespace
Previously the highlighting could fail if it is contained in a region of
a syntax (for example in a markdown code block). As these regions don't
specify the custom highlighting to be contained in them vim won't try
matching them.

This could be fixed by specifying that the highlight match is contained
in ALL regions.