vim: Use matches instead of syntax for special Hi
In some files (mostly latex because of plugin) the custom syntax
elements got overwritten and therefore tabs and trailing white space was
not shown any longer.

Matches shouldn't be interfered by any plugin as it is unlikely that the
plugins use the same ID for the match as I do. Also matches can be
disabled on a window basis. This allows me to disable the highlighting
in e.g. help files as it is of no need there.