Add mpi debug function
The function is capable to attach to local debug_wrapper processes.
Hereby an ORDER is given specifying which process to attach to.

The script is also able to parse a hostfile in the MPI format to attach
to remote processes accordingly. Hereby the number of processes is
continuous over all hosts allowing for easy scripting around the
function.