fish: Cleanup ranger and showDirs
When no ranger was installed the alias executed a `cd` regardless of
that fact. This results in changing directory to home. This was annoying
so here is the fix.

When splitting of a shell with tmux or going in a sub-shell showDirs got
reset. As this value mostly got set for a reason I want to keep it in
children shells.